Reorganized the Preferences screen into tabs, adding helpful text and descriptions for certain items. The Retina option now has a warning attached to try and teach people that there's a severe impact from using it and they don't need it just because they have a retina screen. The preferences are now all Auto-layout setup as well, enabling future localization.
Misc bug fixes:
- Fixed an Autolayout ambiguity that caused console spew in the First Aid pane
- Fixed a misconfigured outlet on the main screen called "dicloseButton" that caused console spew
- Fixed a typo in the NOT_PLAYABLE string
-
