+++ XIV on Mac/SettingsAdvancedController.swift
@@ -1,0 +1,44 @@
+//
+//  SettingsAdvancedController.swift
+//  XIV on Mac
+//
+//  Created by Chris Backas on 4/30/22.
+//
+
+import Cocoa
+
+class SettingsAdvancedController: NSViewController {
+
+    @IBOutlet private var keepPatches: NSButton!
+    
+    @IBOutlet private var esync: NSButton!
+    @IBOutlet private var wineDebugField: NSTextField!
+
+    override func viewDidAppear() {
+        super.viewDidAppear()
+        updateView()
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do view setup here.
+    }
+ 
+    func updateView() {
+        
+        esync.state = Wine.esync ? NSControl.StateValue.on : NSControl.StateValue.off
+        wineDebugField.stringValue = Wine.debug
+        
+        keepPatches.state = Patch.keep ? NSControl.StateValue.on : NSControl.StateValue.off
+    }
+    
+    func saveState() {
+        DispatchQueue.main.async { [self] in
+            Wine.esync = esync.state == NSControl.StateValue.on
+            Wine.debug = wineDebugField.stringValue
+            Patch.keep = keepPatches.state == NSControl.StateValue.on
+        }
+    }
+
+    
+}
